requirements.txt
google-genai
Pillow
openai
anthropic
lunar_python
